+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Fix potential issues in test tasks
 - Add `normalize.css` to available modules to add
 - Fix bug in `addIgnored`
+- Replace Mocha PhantomJS with Karma
+- Remove `test/vendor` files
 
 #### 0.4.1 (November 5, 2013)
 - Update EditorConfig